+++ opportunistic_link/src/request_pointcloud_link_endpoint.cpp
@@ -91,9 +91,9 @@ public:
                 clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &st);
                 sensor_msgs::PointCloud2 cloud = decompressor_.decompress_pointcloud2(res.cloud);
                 clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &et);
-                float secs = (float)(et.tv_sec - st.tv_sec);
-                secs = secs + (float)(et.tv_nsec - st.tv_nsec) / 1000000000.0;
-                float ratio = ((float)cloud.data.size() / (float)res.cloud.compressed_data.size()) * 100.0;
+                double secs = (double)(et.tv_sec - st.tv_sec);
+                secs = secs + (double)(et.tv_nsec - st.tv_nsec) / 1000000000.0;
+                double ratio = ((double)cloud.data.size() / (double)res.cloud.compressed_data.size()) * 100.0;
                 ROS_INFO("Decompression of %f %% took %f seconds", ratio, secs);
                 if (override_timestamps_)
                 {
